catalogue item detail folder refactor
