+++ components/catalogueItem/default-properties/item-properties.comp.tsx
