+++ components/catalogueItem/disclosure/disclosure.comp.tsx
