+++ components/catalogueItem/header/item-detail-header.comp.tsx
