+++ components/catalogueItem/item-detail.comp.tsx
